Avoid using "It" to remove ambiguity
Should not use pronoun "It" in a technical documentation as the sentence is unclear as to what "it " is referring to.
Modified the sentence " It is recommended to skip the first step as it requires creating an account and generating an API key:" to  remove the use of "it". The word "below" is added next to "The first step" to enhance the clarity of the sentence.